Included a js_of_ocaml version of superbol in the .vsix
This is a first step towards fixing #76 (but does not entirely fixes
it).  It avoids the need to install an external superbol executable, and
should work on all platforms.

The bundled superbol-free.bc.js is used if the option `superbol.path` is
set to `null`, which is the new default. Some performance investigation
is needed -- if we find out that the js_of_ocaml version is too slow for
practical use, we can make this not be the default.